IJPL-578 preload kotlin script libraries in advance in kotlin highlighting tests to avoid flaky failures
If the libraries are not loaded, they will be added to the project right from the highlighting pass. This will trigger WA, which cancels current highlighting and restarts it at some point in the future. Test may check the state before the highlighting is restarted (this will be the test failure), ot after it is restarted (test pass).

This commit eliminates the highlighting restart

GitOrigin-RevId: a79d699dca94f5eac2690160ca87c2f576636c28