IDEA-346553 Fix PlatformTestUtil.isBusy for try-again-scrolling
PlatformTestUtil.isBusy relied only on the model for checking
whether any async background stuff is going on with a tree.
However, this didn't work correctly for some async operations
not involving the model. One example is the try-again-scrolling
in TreeUtil, when a tree would scroll to some path and then
try again later to double-check in case path bounds have changed.

Ideally, this try-again-scrolling should be fixed properly,
as it's nothing but a hack. But as it's usually the case with similar
hacks, it's extremely hard to fix it without breaking anything.
And since it seems to be working properly in production,
breaking it by trying to fix some tests doesn't sound like a good idea.

To fix tests, add a client property in TreeUtil that would indicate
that the tree itself is busy doing something that involves
invokeLater operations, even if the model seems to be idle.
Use this new property in PlatformTestUtil.isBusy to check if the
tree itself is busy.

This seems to fix the testAnnoyingScrolling() test for good.
Its flakiness reproduced locally very well before this fix,
and now it seems to have gone completely. Which is expected,
as debugging has clearly shown that this try-again-later scrolling
was indeed the cause: it would scroll to the initial path first,
then the test would wait until the tree is not busy, which wrongly
indicated that it wasn't, and then, 5 milliseconds later, the tree
would scroll again, thus invalidating the captured position.
Now the test properly waits until this second scrolling is done
before going on with its stuff.

GitOrigin-RevId: 9a7ded332677331a7c1460ad1fdae6e3c443da24