IJPL-666 Fix invalidating cached preferred size in DefaultTreeUI
Since DefaultTreeLayoutCache calculates heights lazily for performance
reasons, the cached size in DefaultTreeUI, which is calculated in
updateCachedPreferredSize(), may get outdated. Usually it's not
much of an issue, as something causes it to be invalidated anyway,
for example, moving, resizing, or scrolling the container.

However, in some cases, like the Welcome Screen, the tree is created,
its size is calculated, and then it's just left alone with an invalid size.

To fix this, we introduce a flag in DefaultTreeLayoutCache that is
set every time the preferred size is calculated and cached, and reset
when the cache recomputes the height of some node, and it turns out
to be different. Next time, when we calculate the preferred size,
we check that flag, and if we see that the cache decided to invalidate
the size, we propagate that invalidation to DefaultTreeUI, to force
it to recalculate the size again instead of just using the cached
value which is bound to be invalid now.

GitOrigin-RevId: 777d829d0497f891338f7962266dae295ea8400b