[json] IDEA-346386 Do not use weak keys map in the json objects storage
Mistakenly chosen api was replaced by an ordinary concurrent factory map with strong keys/values for now. Cases when we want to GC already read schema objects are: 1) when a schema is edited manually (very rare) 2) when we don't need a schema anymore (more likely to happen, but really hard to detect). Until we invent a god heuristic, we'll try to avoid letting GC of the schema objects to avoid several computations for the same schema file

GitOrigin-RevId: 28e59675f8af1cedc134467c974d3c3925bd4504