[git-index] fix calling diff in the staging area by double click or enter
Double click and enter handlers try to first call the "Merge" action before invoking diff, in case a node with conflict is selected. "performDumbAwareUpdate" was invoked to update the action, but after 15d32dad64857f15e49b95e5a19af77d34e4131c this does not work anymore. Because of this "Merge" action started to get called even for the non-conflicting nodes. This commit uses utility method instead of an action.

IDEA-346579

GitOrigin-RevId: 7d03e750220a98bce27184c2303562439e6f0e15