IFT-555 Validate task id as simple integer
There is no need in the separate validation rule for it.
Also, this validation rule has side effects: it executed the LessonContext of the lesson to collect the number of tasks. But some for some lessons, it modified the state of the lesson. For example, the code in `OnboardingTourLessonBase.checkUiSettings` modifies the variables of the onboarding lesson, so it caused the problem in the mentioned ticket.

GitOrigin-RevId: d39d464eb0635dcf58842816cfa46fe0e1c4ab82