IDEA-338234 ToolbarFrameHeader: add missing super.removeNotify() call
Because of missing super.removeNotifyCall(), IdeJMenuBar/LinuxIdeMenuBar associated with the ToolbarFrameHeader (created at IdeRootPane.kt:224) did not release its resources properly. So the corresponding entry `Hashtable.Entry(ComponentKeyStrokePair(IdeJMenuBar, Keystroke), IdeFrameImpl(->project))` in KeyboardManager was stuck forever leaking the corresponding project.

The issue was reproducible on Linux and Windows, but not on macOS.

GitOrigin-RevId: 2b99d4d8df271a33debd46c38c5b4c0e39f3bcc3