[gitlab/github] Show create gist and create snippet actions when there's an account or repository (IDEA-343813)
#IDEA-343813 Fixed

Trying to strike a balance between users feeling the creation actions are out-of-context
and users missing the actions from their regular palette of context menu actions.

Now the actions are shown if (the action can be performed and) there
is at least one repository of the required type OR the user has a logged in account.

GitOrigin-RevId: 2b2561254120f695d568de7ff9b1da5ee25267f7