[terminal] IDEA-346817 Do not add custom caret move shortcuts in PowerShell
In Bash and Zsh, Ctrl+A and Ctrl+E shortcuts are moving the caret to the start or end of the line.
But in PowerShell, Home and End shortcuts are doing that. And Ctrl+A is selecting the text instead.
Also move custom shell shortcuts declaration to a single place.

GitOrigin-RevId: 35fb6d75b46117770599e40f45949b50c7de067c