IDEA-346956 IJent: rework error messages written on IJent bootstrap failure
There are some WSL distributions like `rancher-desktop-data` which don't have `uname` inside. IJent can't start in such distributions now.

Before this commit, there was a vague error like "No binary for architecture []", even though the binary has actually existed. Now there's a message "Empty output of `uname`" that indicates on output parsing error.

In addition to a better error message, a part of stderr is attached to the error, so it's possible to find out that `uname` returned nothing because there is no `uname`.

GitOrigin-RevId: 73a0060afe24cae96ca9b4d2b1474d58ffe64dcd