IDEA-346905 Restore the default behavior in RenderingHelper
The logic here is a bit complicated, as it's based on nested negation,
but the point is that for trees that don't satisfy the condition
hsb != null && hsb.isVisible()
the default values were used prior to the recent change.
And these default values are
myShrinkingDisabled = false;
myShrinkingSelectionDisabled = true.

The recent change affected all trees that satisfy the condition
pane.getHorizontalScrollBarPolicy() == ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER.
However, this change affected the defaults as well. Fix by
changing the value for myShrinkingDisabled from
myShrinkingDisabled = !ClientProperty.isTrue(component, SHRINK_LONG_RENDERER)
to
myShrinkingDisabled = ClientProperty.isFalse(component, SHRINK_LONG_RENDERER).
Now the default value, if the property is not set,
is false, matching the old default. For trees that have this
property explicitly set, nothing changes, the new behavior
will still be applied.

GitOrigin-RevId: f4a6728b736f45c035ee5bf3b890ed8b80ab811f