[terminal] IDEA-347062 Introduce separate method for obtaining active command block
It is needed to make possible to create the block lazily after we send the command to a Shell. There will be a period of time when we registered the listeners that is able to request the active block, but it may be not created yet. So, we need to return null in this case. Existing `getLastBlock` is not suitable here, because it is required to check that it is a new block, not the finalized one.

It would be great to close the block right after the command is finished, but it will break if multiple commands are executed at once. So, it should be solved together with this problem.

Also, removed the obsolete comments about 'clear' command. They are not actual, because blocks are removed after the command end now.

GitOrigin-RevId: a19b3f5e1cb1052cedd1f1608e5cd87b4f4feccb