[terminal] IDEA-347062 Create a block for started command lazily
So, if the command is finished fast, we will create a block for the whole finished command. And there will be no visible blinking for fast commands when prompt and the command is painted first, and then output is added in a moment.

(cherry picked from commit 7934506234149f268030ecd40f531118cb9076b3)

IJ-CR-126816

GitOrigin-RevId: c664798616d42cbd6aae442bf7649aeca620d936