RDCT-1083 Context menu does nothing in diff editor
The `BackendActionGroup` is still an empty `DefaultActionGroup` group (to be removed). It is configured in `com.jetbrains.rdclient.ui.diff.UtilKt.setupDiffEditor`. The group can be dropped in the new update mode but let's keep it for a while.

(cherry picked from commit 4976668392263acb0c038c3782632da3d778d882)

IJ-CR-126741

GitOrigin-RevId: 5c6058ad467724a40388ffc64559329004a43870