[vcs-log] use a smaller first step commit count value
The first filtering request to vcs uses FIRST_STEP commit count. Before this change, it had a value of 2000, which seems too high and leads to a delay before the first filtering results appeared. This commit sets the FIRST_STEP value to 100 and moves 2000 to the new SECOND_STEP stage. This can help to speed up showing initial filtering results.

GitOrigin-RevId: 259d3df9d3f3a14656b79388cf1fa2738bdc67d8