IDEA-317864 Create a class to track scroll pane states
To sort out this mess with borders and scrollable content,
we need a tool to keep track of scroll panes and their
scrolled state.

These two classes are designed just for that: one keeps
track of all scroll panes in a given container, the other
is responsible for tracking a given scroll pane and can
be used separately.

GitOrigin-RevId: f3c0c0289612457f2ffb9488ba7715a7c5b35e7f