IDEA-317864 Use the new scroll pane tracker for toolbars and headers
Specifically, in SimpleToolWindowPanel and ToolWindowImpl for
painting the border and the tool window header.

To avoid double borders with ScrollableContentBorder, introduce
client properties that indicate that there's already a border nearby,
so there's no need to paint another one.

Because ScrollableContentBorder can be installed on a different
component than the scroll pane itself, also introduce a weak reference
property to indicate which target component the border is installed
on, so that all checks work properly even if the scroll pane is located
elsewhere.

GitOrigin-RevId: 602998a9346ce2503e56bc9823fb53bc6af8c248