IFT-556 Fix stop action GotIt position in onboarding lessons
`checkToolbarIsShowing` workaround is no more needed because now the toolbars in the `JBTabsImpl` are showing only for the selected tab. This workaround caused the problem because the debug actions toolbar was moved to the left and `x != 0` became false.

(cherry picked from commit 319ef01b448ef024a81357d34093c8cdd5492d26)

IJ-CR-127056

GitOrigin-RevId: 2f8cb5a8739836ab7124c7207663e4ff3701fc0d