IDEA-345583 Capture the data context earlier in refreshScopes
This regression is caused by focus changing by the time
getScopeDescriptors is called from the coroutine.

To fix, we need to capture the data context ASAP.
To keep the API BGT-safe, only do so when the current
thread is the EDT, otherwise we fall back to the previous
dispatch-to-EDT approach.

GitOrigin-RevId: 5ef1ee0f70c6c63c19a52d3e9fb3922405274a6d