GTW-2076 Fix Cannot search for settings that come from host's IDE
To search for settings `SearchableOptionsRegistrar` is used. It can return a list of configurables based on the provided sentence.
To be able to provide the correct result, the `SearchableOptionsRegistrar` stores the mapping from a word to suitable configurable.

These words are either parsed from `searchableOptions.xml` files or provided by `SearchableOptionContributor`s
Files and contributors for host's settings exist only on host's side, therefore, the search is unavailable for them.

To enable search for host's settings, we need to provide the correct mappings (all host's configurables are already synced)
We can do so by initializing `SearchableOptionsRegistrarImpl` on app session creation and sending all processed options to frontend

GitOrigin-RevId: f3ae1c4c106cf6d633f80d5e923d481605d5250a