[rdct tests] don't delete tests plugin while clearing config directory (RDCT-773)
In the new modular scheme to run JetBrains Client in tests, the test code is included in a separate plugin, so we need to ensure that the plugin isn't removed while preparing the test. Also, we cannot configure the test to start with clean config because this will remove the plugin as well. It's enough that the config directory exists to ensure that config won't be imported from other places.

GitOrigin-RevId: c19221dee0139b1ef0e4ee06d787413eb3fd7c36