[build scripts] include runtime module descriptors to cross-platform distribution (IJPL-761)
Currently, the only differences between module-descriptors.jar for different OS are caused by platform-specific plugins (e.g. emojipicker) which modules are present in one repository and absent in others. So to get a repository for the cross-platform distribution, it's enough to find descriptors which are present in all repositories. Their dependencies are assumed to be the same, and resource paths are intersected.

GitOrigin-RevId: 85d4df526204e045aa378a623dee32d88b501606