IDEA-270009 Fix tool window switcher when too many windows
The issue is caused by having too many tool windows, breaking
the logic in getIndexShortcut(), making it return a string consisting
of more than one digit, e.g., for 36 it returns 10 (which is 36
in the "thirty-heximal" notation). And then it breaks because 10
isn't a valid key, obviously.

Fix by making getIndexShortcut() return null instead and then
handling it carefully wherever it's called.

GitOrigin-RevId: 075626bbeec4e98ae4cb4d8a0aea4df426a5e63d