[ruby] fix RbsParserTest to only test diagnostics against a single parsing pass
There were issues where `configureByFile` could trigger multiple parsing passes of the created file. For example, the file would be set up with different content (empty vs. non-empty). The majority of the time we still only had one pass over the non-empty file, but sometimes due to a race condition we would get multiple passes, one for the empty file, and one for the non-empty file.

This change ensures that it doesn't matter how many passes happen during `configureByFile`, and instead we only run the diagnostics on a single parse.

GitOrigin-RevId: ad1887db59774504ae5b98980070a7604218e1f7