IDEA-295653 Use last mouse press point to check for gutter dragging
When we check whether we're about to drag a gutter icon,
we need to use the point where the mouse was initially pressed,
not its current location. These two can be quite different,
especially on Linux where dragging seems to start with
significant delay (likely to protect against accidental drags).

Convert the point from the editor component to the gutter
coordinate space to ensure getGutterRenderer works
properly (not obvious from its code).

GitOrigin-RevId: 89116c0842723c6db9b71dab6a87dedfb74451b2