KTIJ-28887 [kotlin] Take the type of companion objects into account when completing classifiers
We used to use the type information only for regular `object`s and enum
entries.
It's only natural to also account for the type of companion objects,
since they also can be used as values through the containing class name.

A prominent use-case for that behavior is `androidx.compose.ui.Modifier`
in the Compose UI framework.

NB1: Note that since companion objects are not necessarily visible on the completion
side (even when the containing class is visible), we perform the explicit
visibility check on them so that the types of non-visible companion objects'
instances are ignored.

NB2: There is already a `ComposeModifierCompletionContributor` class
which adds the extension functions on `Modifier` type to the completion and
pulls them to the top in Compose projects when Android Plugin is enabled.
This fix does not interfere with that contributor, since when somebody
tries to complete for `Modifier`, they would probably start typing
`Mod` with a capital letter.
Since the contributor uses prefix matcher to filter the proposed extensions
by their names, the capital letter in the `Mod` prefix will (most probably) filter
all the lower-case extension functions proposed by this contributor
(see `com.android.tools.compose.code.completion.ComposeModifierCompletionContributor.getExtensionFunctionsForModifier`)

This behavior is currently not fully supported in K2, it will be done separately
later (see KTIJ-28961).

^KTIJ-28887 Fixed

GitOrigin-RevId: 97759ef903df90e9f8a6095d80ec91b854bf8f29