[amper] Add resolution anchors for the Amper parser
Amper libraries depend on published IntelliJ modules. However, when using them in the plugin, the modules are excluded from the transitive closure, and instead we depend on the sources of these modules. That leads to the resolution problems in Kotlin. To solve that, we use the same solution as Kotlin by providing anchor to the module containing the desired classpath of the Amper library.

GitOrigin-RevId: beb7f840fc96acc83aec8151ca5b0ff116f03305