IDEA-336296 Fix editor gutter drag selection resetting
The issue was caused by the fact that we have two selection
drag modes: dragging on the line numbers area and dragging
on other areas. They work a bit differently, and when we switch
from one mode to another, the selection is reset.

Specifically, the selectLineAtCaret(true) call that happens
when the line number area drag selection starts, resets the
selection to just one line.

To avoid this, introduce a boolean flag that's reset
on mouse press and set whenever we start drag selection
in any mode. Then check this flag, and if it's already set,
just don't reset the selection again.

GitOrigin-RevId: 63e65b4bc7e3ae4cadc0019db53f9066c5da1f63