[github] Fix adding a comment marks github PR as unread
Shouldn't be the case if you're the user adding the comment (from IDE).
This also simplifies the marking of last seen a bit:
when the user updates any PR from IDE, the time of the update is recorded as the last seen date.

(cherry picked from commit 74483e7671368ce84a3318ef16af07866357e272)

IJ-CR-128256

GitOrigin-RevId: 989eb8714a3dc6e6a34b956d9c633ac12b8b391b