[platform] IDEA-343397: An experiment: attach children only to a registered disposable parent
This is an attempt to fix IDEA-343397. In the constructor of DomManagerImpl, the children get attached to a disposable before it gets registered as a child of project container. I conjecture that an exception in the constructor leads to DomManagerImpl attached to the ROOT tree instead of the project, which leads to incorrect dispose in tests. If the conjecture is true (i.e. there are no failing tests after this change), then we should deal with this problem on a platform level

GitOrigin-RevId: 26c3ae7a2d0627344c544e724239d8fdd8799024