IDEA-348709 Create TreeBulkExpansionListener
Now that we have bulk expand/collapse tree operations,
it's sometimes necessary to get callbacks not just when
another node has collapsed/expanded, but when the entire
bulk operation starts/finishes.

One such example is TreeTableModelAdapter, which
tries to save and restore selection. If this happens
during a bulk collapse operation, it causes collapsed
nodes to re-expand. We'll fix this using this new API.

GitOrigin-RevId: f7eb6c7eff651f07a887f8da807a2bd275e88caa