[kotlin-dfa] UsefulNull: extract x == null to variable
It looks like, due to subtle changes in type system, we won't be able to issue 'x is always null' warnings in K2 when 'x' is smart-casted to null. Arguably, such a warning is unnecessary and could be covered by some simpler inspection.
Preparatory step for KTIJ-27359 K2 IDE: migrate KotlinConstantConditionsInspection to K2

GitOrigin-RevId: bf4221c166ebe199804acaaa7202203a6c1099e4