[vcs-log-graph] use all branch heads for the graph layout
Head commits are used to calculate layout indexes, which are used for sorting nodes and edges and coloring. However, before this commit only graph heads, and not branch heads were used. Because of this, some commits belonging to important branches were not laid out or colored properly, just because the important branch head was contained in some other less important branch.

GitOrigin-RevId: 1cd228ff6107d0ca03d5f3172f9940e0d55d9e4f