IDEA-347891 Don't call setBorder from getPreferredSize
This is an anti-pattern and can cause infinite repaint loops
(like this issue) and other problems.

The perfect fix would be to just update the preferred size
once when the state changes and cache it somewhere,
but unfortunately the size depends on the state outside
the component, for example, the number of tabs in
the tool window. It's impossible to reliably track updates
to the whole state from within the component.

So instead of a proper fix, let's work around the issue
by not calling setBorder, but instead setting up a border
with mutable insets once and then just updating those
insets. This way, we at least don't have any hidden repaint()
calls in getPreferredSize().

(cherry picked from commit fd31eba44f2c90de503f3b22210d007af09c7319)

IJ-CR-128482

GitOrigin-RevId: 0a7f365185bfbbb2b03887226fc167c9542c6c34