GTW-7868 Fix Recent project window stuck after opening a project in the same window
Opening project in the same window causes the current project to be closed.
With the project closing, all opened popups should be closed too.

However, before closing `AbstractPopup` calls `canClose` method to check whether the popup can be closed. The method itself invokes `cancelCallback` which can also signal that the popup cannot be closed.

If the method returns false, the popup will not be hidden, but the `AbstractPopup` will be disposed which prevents all further possibilities of hiding the popup.

To avoid stuck windows, `AbstractPopup` should hide the popup window on dispose even if the `canClose` or `cancelCallback` signal over-wise.
We can use `myDisposed` flag as an indicator that we're in the process of disposing the popup

GitOrigin-RevId: ab95f0e805602363f11552544330c80ca222969b