(IJPL-766) Preprocess query and item text for embedding generation in the same way for standard and semantic items
Previously, the cosine similarity with a search query could be different for the same items found by standard contributor and for items found by semantic contributor. That was because in the latter case, all labels were split into tokens via a regular expression. To match this behavior for standard items, I perform the same text preprocessing in feature provider.

GitOrigin-RevId: d1c0ca47b0454332d77992bada1e574fa644c8fe