IDEA-253911 Catch exceptions in the scope model
Since scope functionality is used in some important places,
such as Find Usages, and a failure to create a single scope
may mean that the scope model is never initialized,
we need to be extra careful with extensions here.

Because extensions can contain any code, the safest bet
here is to catch and log all exceptions, and simply skip
such faulty providers.

GitOrigin-RevId: 081ae0a3a267eb869e828a613ce06ad3c2e9673e