[pycharm] get rid of excessive dependencies in intellij.pycharm.community module (RDCT-1034)
This module needs to be added to the frontend variant of PyCharm (JetBrains Client) for RDCT-1034, and dependencies on modules which aren't included to JetBrains Client will cause problems. So they are removed from the module, and missing dependencies are added to intellij.pycharm.community.main and intellij.python.community.tests modules instead.

GitOrigin-RevId: 8818588e7f6a55ace63ce48edd63bb88938d4a16