IDEA-348467 Don't start dragging in editor unless mouse moved far away enough
It may be very hard to perform a click, e.g., to set a breakpoint,
because if the mouse moves even a little bit, it's interpreted as a drag
instead.

Fix by introducing some sensitivity and only start dragging when
the mouse moves far away enough. Introduce the editor.drag.sensitivity
registry key for this and set its value to 5 for now, which is an
initial estimate made by simply experimenting in a single environment
for a few minutes.

For now, only apply this sensitivity to the gutter, as it seems to be
the most problematic place, and applying this editor-wide breaks
a lot of editor tests that seem to rely on artificial micro-drags.
Those are, of course, fixable, but considering the amount of effort
and that we don't really need it at this point, limiting this to
the gutter seems like the best middle ground for now.

GitOrigin-RevId: 46b20a89c3f3127402ad4ab5b6de3a2cf448ca07