[debugger] don't update breakpoints UI if project is disposed
There is a problem with GoColorSchemeTest.testDarcula() that fails with "already disposed" exception. The exception indicates that during tearDown() stage it flushes NonBlockingReadAction queue which has the task that invokes BreakpointManager.getBreakpoints() method after a projects is disposed. This method is called in BreakpointsManager.updateBreakpointsUI() in a non blocking read action without expiration condition.
We fix this issue by explicitly expiring non-blocking read action if the project is already disposed.
Fixes GO-16412.

GitOrigin-RevId: dec2653d590cf30f410f02207eafaa9cfa40de26