[devkit] provide highlighting and completion for product-modules.xml (IDEA-348550)
DOM model is added for product-modules.xml, and ProductModulesXmlDomInspection is introduced to report errors in it. The inspection is marked as internal and disabled by default, because currently it makes sense for the 'intellij' project only.
References to modules in product-modules.xml use IntellijModuleSymbol as the target, but for now it isn't integrated in Symbol API, and its data is used directly to resolve references to modules in IntellijModuleConverter.

GitOrigin-RevId: 98f264fdad920b518125c6dfcff794d31778c1c6