IJPL-578 drop SmartModeSchedulerStartupActivity
It does not wait until startup activities are finished: all the startup activities are running in parallel in prod. This activity reliably waits until startup activities are started. Remove the code to avoid confusion.

In addition, in tests activities do run one after another, so we may get a deadlock when some startup activity waits for smart mode which will never start because SmartModeSchedulerStartupActivity occurs further in the queue.

GitOrigin-RevId: b387aa010ce0e3ca04a23fac5312526ce4ea4052