Introduce LighterASTSyntaxTreeBuilderBackedNode for efficient access to underlying text
This commit introduces LighterASTSyntaxTreeBuilderBackedNode which represents a lighter AST non-leaf node that is backed by a syntax tree builder and provides efficient access to its underlying text without the need to (recursively) traverse its children. In addition, LightTreeUtil.toFilteredString is adapted to take this type of nodes into account and directly return the text.

GitOrigin-RevId: 6425c5a78013584007ebd1abf66468c2820297d4