IDEA-253911 Fix ScopeService job cancellation in tests
There are two issues here. The first, some tests temporarily dispose
a project without actually disposing it, and that led to the scope service
doing something with a disposed project, throwing various exceptions.
The second, the next test may reuse the same service again,
so the service somehow needs to be disposed when the previous
test disposes the project, but then it should be not disposed again.

Fix both issues by removing that ugly hack for temporarily disposed
projects and instead make the scope model properly disposable.
Now, when the scope chooser combo is disposed, the model
is disposed too, cancelling its jobs, but not cancelling the service
because the model now gets a child scope (as it probably should've
to begin with).

GitOrigin-RevId: d82bfa2c2a4d3ded7cc0ed081870eded879282fe