Adapt to compiler plugin changes.
The parcelize compiler plugin has been updated to accept additional
annotations as a plugin option. There additional annotations work
the same as the `Parcelize` annotation - they trigger processing
of the element they are attached to.

Because of this new options, the list of parcelize triggering
annotations need to be passed when configuring the parcelize
plugin. This change does that for the IDE.

GitOrigin-RevId: 8627192f70b7bffb39c0137965a143fc57a995b4