IDEA-348158 IJent+WSL: copy IJent binary into container via 9P
The previous approach with `cat` locally and `head -c $binarySize > $binaryPath` remotely turned out to be unreliable. The `head` call may have frozen infinitely. By the moment of writing this message, it was not clear why it had been failing.

This commit allows uploading an IJent binary in a different way. The interface is tied to the current implementation and aims to fixing particular bugs with uploading over WSL. A better API is to be provided later.

(cherry picked from commit dd5cf2a54193ea9054dbadc6d095ecff8b045ff8)

IJ-CR-128716

GitOrigin-RevId: 2a889b7bd53b77768a648573ba39232ac3828a2c