[platform] always use '-n' argument for 'open' command when restarting the IDE on macOS (RDCT-1176)
On macOS, the full IDE and its frontend variant (ex JetBrains Client) are started from the same application bundle. So if they are running concurrently, it's necessary to use '-n' argument for 'open' command, otherwise the OS will just activate the running process instead of launching a new one. In regular cases, this argument does nothing because the IDE process finishes before the command is executed.

GitOrigin-RevId: 281fa311331bd327e8440ad8ee5bd7f527c69eb4