[diff] improve combined diff drawing performance by avoiding clipping painting
There are 2 ways of truly painting only a part of a component - clipping and masking. Both of them have abysmal performance, which significantly impacts FPS when scrolling combined diff.
To work around that we go for a hack - we clip the corners by filling the area around the rounding with background color.

GitOrigin-RevId: 771306a086eb4205b34394d54ef2b591130c5370