[core][ui][IDEA-348959] replace Disposable with CheckedDisposable in AbstractTreeModel
The current implementation of `Disposer` uses `ObjectTree` to store references to the `Disposable` registered in the application. As a result, a hard link to `AbstractTreeModel` is retained even after `dispose` is called. `CheckedDisposable` prevents the source object reference from being stored in `ObjectTree`, so Garbage Collector can collect `AbstractTreeModel` much faster.

GitOrigin-RevId: ed6dd692cbcb984b443a9ce8453474f3e779dbf5