[editor-inlay] use editor viewport size to layout inlays instead of visible bounds
Using visible area of a viewport for size reference caused inlays to be resized when the editor was obscured by some other component or not visible. This caused issues when an editor was placed in another scrollpane (ex combined diff), because when it's scrolled away the inlay was resized to zero width and not all components can handle this (definitely not JEditorPane).

#IDEA-349311 Fixed

(cherry picked from commit 5bab239dce9b57027c644ad1cbd715b8a68e8320)

IJ-CR-134394

GitOrigin-RevId: c4124064525ab79e51ca3b278663b9b512243041