[module-based loader] provide a way to skip some root and plugin modules when including content in product-modules.xml (IJPL-128)
Sometimes we want to include content of product-modules.xml from a base product, but skip some modules from it (e.g., intellij.cwm.guest includes 'performanceTesting' plugin by default, but required modules are missing in some IDEs, so it won't work in them). This commit introduces 'without-module' subtag in 'include' tag which allows to still reuse content of the base product in such cases and explicitly mark modules which should be skipped.

GitOrigin-RevId: 57970894cda9de9bb46295b5a41d9d0d833a3967