GTW-7963, RDCT-1030 Remove `ActionRemoteBehaviorSpecification.Frontend` from actions which should be duplicated in RD
These actions can be executed both on backend and frontend (they're a part of `UnattendedBackendDelegatingActionIdsProvider.DUPLICATED_ACTION_IDS`)
`Frontend` marker is used to make them frontend-only in CWM scenario.

However, in the new actions mode, this doesn't properly work - backend's actions cannot be found in the main menu. Because now the marker interface is also being checked on frontend and the actions are considered to be frontend, therefore, backend's version shouldn't be shown.

To mark actions as frontend-only for cwm `CWM_BLACKLISTED_ACTIONS` can be used - it'll prevent backend's version from creation and won't affect RD

GitOrigin-RevId: f3a93e75849d8ae48a750a3071cbc58d8c72ad27