[vcs-log-graph] improve edge comparison to avoid intersections
This commit fixes the comparison between edges originating from the same node. To compare such edges, an end of the shorter edge is compared with the longer edge. If the end node of the shorter edge is positioned on the left side, the shorter edge is also positioned on the left side, and the same principle applies if the end node is on the right side. This allows avoiding unnecessary edge intersections.

GitOrigin-RevId: 944212e45f6fd2b975eda3acbad24fa898a0a29a