Build the updater-full.jar with Gradle
With IntelliJ moving to build the updater jar with Gradle, we need to
also. Otherwise, we get a corrupt updater jar after d3b1e47df7.

As part of this change we need to specify the JAVA_HOME for our build
server to invoke Gradle so we use the one from prebuilts. We could use
the JDK IntelliJ uses for invoking the JPS build, but it is stored in
a directory based off the specific version - so there isn't a good way
to make our Gradle updater jar build resilient to the version changing.

Fixes: 329723843
Change-Id: I90b01ea8160375dca37919b8b4c081f9fe8f9897