BAZEL-888 [Auto-Link] fix: use the project base path as a default project root
The guessProjectDir returns the content root of any project module. Some build systems don't create the project root's module. In their case, the suggested project directory points to a random directory in the project. That is incorrect for the basic unlinked project scanner. All advanced cases should be covered by the ProjectRoots scanner.

GitOrigin-RevId: f5c147299e6aff8170d354dcd16f83126e43676e