PY-64402: Only create `main.py` if "create welcome script" checkbox set for the Jupyter project.
We used to create `main.py` in any case, but must do that with checkbox.

This code is not elegant due to boilerplate and possible inconsistency (generator might ask for the welcome script and ignore it), but this is the only API for now.

(cherry picked from commit 9ca3ab29b763c089c0f449045f1418d8f9a96896)

IJ-CR-129683

GitOrigin-RevId: 10e937957655e51f10a423ce9bed1d528258dc1c