[collab/git4idea] Fix editor state restoring unnecessarily on opening closed diff
When opening a closed diff, we already restore the exact selection
of a block diff manually, so we need to make sure that is not also restored
from file history.

By excluding this information from FULL-level history, it will not be a
part of the state that's restored when a new editor is opened.

GitOrigin-RevId: a579b5150d5f4529b52a381e8141030d1676e31f