IJPL-928 Mark EditorImpl.RepaintCursorCommand as context-aware
Scheduling a runnable using scheduleWithFixedDelay wraps
it into a ContextRunnable under the hood, which captures the context,
which sometimes contains a reference to the project, creating a leak,
because it's a global app-wide thing.

Since this thing is pure UI code and doesn't require any context,
we mark it as context-aware to avoid capturing the context.

GitOrigin-RevId: 23240d901b6bae541b103563399a8030d5e4f1d6