IDEA-317864 Use ScrollableContentBorder in ProblemsViewPanel
To account for orientation changes, extract the code setting
the toolbar border into a function and do everything there.
It's called from the contstructor and when the orientation changes.

For the old UI it does nothing, so nothing is updated. It's the existing
behavior, so we leave it alone, even though it looks wrong, but that's
a separate issue.

GitOrigin-RevId: 7df540d66fb2d46b3b8cf880335a74fca021b089