[terminal] IJPL-101428 Extract right prompt when prompt is represented by a single escape sequence
It is the case for Bash and PowerShell. There is no separate variable for the right prompt, so both parts of the prompt are in the single escape sequence.
The idea is to split it by the cursor. If there is something meaningful after the cursor - consider it as a right prompt.

GitOrigin-RevId: 6d6a694c3db31478c83de97ffd831c85043d3fc7