[gradle][groovy] IDEA-257715 move delegate determining for Project#artifacts(...) closure into GradleDelegatesToProvider
... from GradleArtifactsContributor. The new logic in GradleDelegatesToProvider allows determining a delegate for a Closure if receiver class has overloaded method with Action<? super T>. In that case, the delegate will be a class substituted into T. It covers some code implemented in contributors, called from GradleDelegatesToProvider. Including GradleArtifactsContributor. That's why I removed a case for the Project#artifacts(...) method from there.

For a case when Project#artifacts(...) method is called, this contributor not only determines a delegate, but also wraps it with GradleProjectAwareType. It allows GradleArtifactHandlerContributor adding some extra methods with names of configurations during the resolving. Since GradleDelegatesToProvider becomes responsible for this case, I also added a replaceWithProjectAwareType method. It adds GradleProjectAwareType wrapping for a delegate if needed. Unfortunately, I did not find a way to put a resolved delegate in wrapping. I recreate a type with the same class name. Because GradleProjectAwareType#setType method expects PsiClassType as an argument. The resolved delegate is a PsiType, its ancestor. So, it could not be used as is.

GitOrigin-RevId: 0914d6bfe264ed8989d910b4391729b57025dd45