IDEA-317864 Add ScrollableContentBorder implementation for complicated cases
Track scroll panes automatically as they're added and add borders if a scroll pane
touching the edge is scrolled. Needed for complicated tool windows like Build,
where it's hard to figure out the exact scroll panes to track, because they are added
to a splitter, which may or may not have other components and it may change
at any moment.

So instead track scroll panes and check their location to see if scrolling should add
a border.

GitOrigin-RevId: 1dae2bf8b5ce43beb54de84a59a58e78721fb3eb