IJPL-935 Honor visibleRowCount when computing preferred size in DefaultTreeUI
It's needed when a tree is located in a small viewport,
and we need to increase the viewport size, but first we need
to compute the preferred size of the tree. Similarly how
a fallback is already implemented for the case when the
paint bounds are completely empty, we now use more rows
than actually visible in the viewport if the tree specifies
its own visible row count.

GitOrigin-RevId: 5c940d47417e5d1b879ef8ac60b2609e27b622ed